src/entities/Profile/model/services/fetchProfileData/fetchProfileData.ts
import { createAsyncThunk } from '@reduxjs/toolkit'
import i18n from 'i18next'
import { type ThunkConfig } from 'app/providers/StoreProvider'
import { type Profile } from '../../types/profile'

export const fetchProfileData = createAsyncThunk<Profile, void, ThunkConfig<string>>(
  'profile/fetchProfileData',
  async (_, thunkAPI) => {
    const { extra, rejectWithValue } = thunkAPI
    try {
      const response = await extra.api.get<Profile>('/profile')

      if (!response.data) {
        throw new Error()
      }

      return response.data
    } catch (e) {
      return rejectWithValue(i18n.t('Произошла непредвиденная ошибка'))
    }
  },
)
